+++ src/nectarchain/trr_verification_package/styles/gui_style.qss
